+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 * Client#dimensions no longer accept a dimension code. Use Client#dimension
   instead.
+* Client#dimension (resp. Client#dimensions) now return instances of the
+  Klaro::Client::Dimension (resp. Klaro::Client::Dimensions) classes, no
+  longer of ruby Hash.
 
 # 0.3.0 - 2020-01-06